formatting, remove unneeded cast
git-svn-id: http://svn.enlightenment.org/svn/e/trunk@35779 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33

+++ ewl/src/lib/ewl_freebox_mvc.c
@@ -163,10 +163,10 @@ ewl_freebox_mvc_cb_configure(Ewl_Widget *w, void *ev __UNUSED__,
         if ((!ewl_mvc_dirty_get(EWL_MVC(fb_mvc))) 
                         || !model || !view || !mvc_data)
                 DRETURN(DLEVEL_STABLE);
-        
+
         ewl_container_reset(EWL_CONTAINER(fb_mvc));
-	 count = (unsigned int)model->count(mvc_data);
-	 for (i = 0; i < count; i++)
+        count = model->count(mvc_data);
+        for (i = 0; i < count; i++)
         {
                 Ewl_Widget *o, *cell;
